moving SearcherFile class to model package
